+++ src/lib/preload-libc/gen_syscall_wrappers_c.py
@@ -64,6 +64,7 @@ skip.update([
     'fork',
     'exit',
     'getcwd',
+    'getrlimit',
     'mq_notify',
     'mq_open',
     'poll',
@@ -75,6 +76,7 @@ skip.update([
     'sched_getaffinity',
     'sched_setaffinity',
     'select',
+    'setrlimit',
     'setuid',
     'sigaction',
     'sigprocmask',
@@ -212,7 +214,10 @@ for line in data.splitlines():
     if len(parts) < 4 or '#' in parts[0]:
         continue
 
-    num, abi, name, entry = parts
+    # Most lines contain only these 4 fields.
+    # Some contain additional trailing fields we don't care about
+    # ("-  noreturn"), which we drop.
+    num, abi, name, entry = parts[:4]
 
     # ignore the x32-specific abi, since shadow only supports x86_64
     if 'x32' in abi: