+++ src/main/host/syscall/handler/resource.rs
@@ -1,39 +1,80 @@
 use linux_api::errno::Errno;
 use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 
+use crate::host::process::ProcessId;
 use crate::host::syscall::handler::{SyscallContext, SyscallHandler};
 use crate::host::syscall::types::SyscallError;
 
 impl SyscallHandler {
+    fn prlimit64_impl(
+        ctx: &mut SyscallContext,
+        pid: linux_api::posix_types::kernel_pid_t,
+        resource: std::ffi::c_uint,
+        new_rlim: Option<&linux_api::resource::rlimit64>,
+        old_rlim: Option<&mut linux_api::resource::rlimit64>,
+    ) -> Result<(), SyscallError> {
+        let Ok(resource) = linux_api::resource::Resource::try_from(resource) else {
+            return Err(Errno::EINVAL.into());
+        };
+
+        if pid == 0 {
+            // process is calling prlimit on itself
+            ctx.objs.process.prlimit64(resource, new_rlim, old_rlim)?;
+        } else {
+            // calling on another process
+            let Ok(id) = ProcessId::try_from(pid) else {
+                return Err(Errno::ESRCH.into());
+            };
+            let Some(process) = ctx.objs.host.process_borrow(id) else {
+                return Err(Errno::ESRCH.into());
+            };
+            process
+                .borrow(ctx.objs.host.root())
+                .prlimit64(resource, new_rlim, old_rlim)?;
+        }
+        Ok(())
+    }
+
     log_syscall!(
         prlimit64,
         /* rv */ std::ffi::c_int,
         /* pid */ linux_api::posix_types::kernel_pid_t,
         /* resource */ linux_api::resource::Resource,
-        /* new_rlim */ *const std::ffi::c_void,
-        /* old_rlim */ *const std::ffi::c_void,
+        /* new_rlim */ *const linux_api::resource::rlimit64,
+        /* old_rlim */ *const linux_api::resource::rlimit64,
     );
     pub fn prlimit64(
-        _ctx: &mut SyscallContext,
+        ctx: &mut SyscallContext,
         pid: linux_api::posix_types::kernel_pid_t,
         resource: std::ffi::c_uint,
-        _new_rlim: ForeignPtr<()>,
-        _old_rlim: ForeignPtr<()>,
+        new_rlim_ptr: ForeignPtr<linux_api::resource::rlimit64>,
+        old_rlim_ptr: ForeignPtr<linux_api::resource::rlimit64>,
     ) -> Result<(), SyscallError> {
-        log::trace!("prlimit64 called on pid {pid} for resource {resource}");
-
-        // TODO: For determinism, we may want to enforce static limits for certain resources, like
-        // RLIMIT_NOFILE. Some applications like Tor will change behavior depending on these limits.
-
-        if pid == 0 {
-            // process is calling prlimit on itself
-            Err(SyscallError::Native)
+        let new_rlim = if new_rlim_ptr.is_null() {
+            None
         } else {
-            // TODO: We do not currently support adjusting other processes limits. To support it, we
-            // just need to find the native pid associated with pid, and call prlimit on the native
-            // pid instead.
-            Err(Errno::EOPNOTSUPP.into())
+            Some(&ctx.objs.process.memory_borrow().read(new_rlim_ptr)?)
+        };
+
+        let mut old_rlim = if old_rlim_ptr.is_null() {
+            None
+        } else {
+            Some(&mut linux_api::resource::rlimit64 {
+                rlim_cur: 0,
+                rlim_max: 0,
+            })
+        };
+
+        SyscallHandler::prlimit64_impl(ctx, pid, resource, new_rlim, old_rlim.as_deref_mut())?;
+
+        if !old_rlim_ptr.is_null() {
+            ctx.objs
+                .process
+                .memory_borrow_mut()
+                .write(old_rlim_ptr, old_rlim.unwrap())?;
         }
+
+        Ok(())
     }
 
     log_syscall!(
@@ -43,11 +84,24 @@ impl SyscallHandler {
         /* rlimit */ *const linux_api::resource::rlimit,
     );
     pub fn getrlimit(
-        _ctx: &mut SyscallContext,
-        _resource: std::ffi::c_uint,
-        _rlimit: ForeignPtr<linux_api::resource::rlimit>,
+        ctx: &mut SyscallContext,
+        resource: std::ffi::c_uint,
+        rlimit_ptr: ForeignPtr<linux_api::resource::rlimit>,
     ) -> Result<(), SyscallError> {
-        Err(SyscallError::Native)
+        let mut rlim_val64 = linux_api::resource::rlimit64 {
+            rlim_cur: 0,
+            rlim_max: 0,
+        };
+        SyscallHandler::prlimit64_impl(ctx, 0, resource, None, Some(&mut rlim_val64))?;
+        let rlim_val = linux_api::resource::rlimit {
+            rlim_cur: rlim_val64.rlim_cur,
+            rlim_max: rlim_val64.rlim_max,
+        };
+        ctx.objs
+            .process
+            .memory_borrow_mut()
+            .write(rlimit_ptr, &rlim_val)?;
+        Ok(())
     }
 
     log_syscall!(
@@ -57,10 +111,16 @@ impl SyscallHandler {
         /* rlimit */ *const linux_api::resource::rlimit,
     );
     pub fn setrlimit(
-        _ctx: &mut SyscallContext,
-        _resource: std::ffi::c_uint,
-        _rlimit: ForeignPtr<linux_api::resource::rlimit>,
+        ctx: &mut SyscallContext,
+        resource: std::ffi::c_uint,
+        rlimit: ForeignPtr<linux_api::resource::rlimit>,
     ) -> Result<(), SyscallError> {
-        Err(SyscallError::Native)
+        let rlim_val = ctx.objs.process.memory_borrow().read(rlimit)?;
+        let rlim_val = linux_api::resource::rlimit64 {
+            rlim_cur: rlim_val.rlim_cur,
+            rlim_max: rlim_val.rlim_max,
+        };
+        SyscallHandler::prlimit64_impl(ctx, 0, resource, Some(&rlim_val), None)?;
+        Ok(())
     }
 }